# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Mentoring System REST API
  version: 0.0.1
paths:
...
  /api/v1/mentoring-requests/from-me:
    get:
      tags:
        - MentoringRequest
      summary: Returns a page with user's mentoring requests
      parameters:
        - in: path
          name: status
          required: false
          example: pending
        - in: path
          name: page
          required: false
          example: 2
      responses:
        '200':
          description: A JSON map with page of user's mentoring requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: object
                    properties:
                      content:
                        type: object
                        properties:
                          id:
                            type: integer
                            format: int64
                            example: 1
                          from:
                            type: object
                            properties:
                              id:
                                type: integer
                                format: int64
                                example: 123
                              name:
                                type: string
                                example: "Name"
                                minLength: 2
                                maxLength: 50
                              surname:
                                type: string
                                example: "Surname"
                                minLength: 2
                                maxLength: 50
                              email:
                                type: string
                                example: "name.surname@email.com"
                              status:
                                type: string
                                enum:
                                  - NEEDS_INFORMATION
                                  - ACTIVE
                              createdAt:
                                type: string
                                example: "21.04.2023"
                              role:
                                type: string
                                enum:
                                  - MENTOR
                                  - MENTEE
                              questionnaire:
                                type: object
                                properties:
                                  about:
                                    type: string
                                    minLength: 250
                                    maxLength: 1024
                                    example: "Some information about user"
                                  skills:
                                    type: array
                                    items:
                                      example: "Skill"
                                  companies:
                                    type: array
                                    items:
                                      example: "Company"
                                  rank:
                                    type: string
                                    example: "Rank"
                                  specialization:
                                    type: string
                                    example: "Specialization"
                                  hoursPerWeek:
                                    type: integer
                                    format: int32
                                    example: 20
                                    minimum: 1
                                    maximum: 168
                                  profilePhotoUrl:
                                    type: string
                                    example: "https://url.com/profile-photo.jpg"
                          to:
                            type: object
                            properties:
                              id:
                                type: integer
                                format: int64
                                example: 123
                              name:
                                type: string
                                example: "Name"
                                minLength: 2
                                maxLength: 50
                              surname:
                                type: string
                                example: "Surname"
                                minLength: 2
                                maxLength: 50
                              email:
                                type: string
                                example: "name.surname@email.com"
                              status:
                                type: string
                                enum:
                                  - NEEDS_INFORMATION
                                  - ACTIVE
                              createdAt:
                                type: string
                                example: "21.04.2023"
                              role:
                                type: string
                                enum:
                                  - MENTOR
                                  - MENTEE
                              questionnaire:
                                type: object
                                properties:
                                  about:
                                    type: string
                                    minLength: 250
                                    maxLength: 1024
                                    example: "Some information about user"
                                  skills:
                                    type: array
                                    items:
                                      example: "Skill"
                                  companies:
                                    type: array
                                    items:
                                      example: "Company"
                                  rank:
                                    type: string
                                    example: "Rank"
                                  specialization:
                                    type: string
                                    example: "Specialization"
                                  hoursPerWeek:
                                    type: integer
                                    format: int32
                                    example: 20
                                    minimum: 1
                                    maximum: 168
                                  profilePhotoUrl:
                                    type: string
                                    example: "https://url.com/profile-photo.jpg"
                          status:
                            type: string
                            enum:
                              - PENDING
                              - REJECTED
                              - ACCEPTED
                          createdAt:
                            type: string
                            example: "21.04.2023"
                          updatedAt:
                            type: string
                            example: "21.04.2023"
                      currentPageNumber:
                        type: integer
                        format: int32
                        example: 1
                      totalPages:
                        type: integer
                        format: int32
                        example: 2
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "about:blank"
                  title:
                    type: string
                    example: "Bad Request"
                  status:
                    type: integer
                    format: int32
                    example: 400
                  detail:
                    type: string
                    example: "Номер сторінки не може бути менше 1."
                  instance:
                    type: string
                    example: "/api/v1/mentoring-requests/from-me"
      security:
        - jwt_token: [ ]
  /api/v1/mentoring-requests/to-me:
    get:
      tags:
        - MentoringRequest
      summary: Returns a page with mentoring requests to user
      parameters:
        - in: path
          name: status
          required: false
          example: pending
        - in: path
          name: page
          required: false
          example: 2
      responses:
        '200':
          description: A JSON map with page of mentoring requests to user
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: object
                    properties:
                      content:
                        type: object
                        properties:
                          id:
                            type: integer
                            format: int64
                            example: 1
                          from:
                            type: object
                            properties:
                              id:
                                type: integer
                                format: int64
                                example: 123
                              name:
                                type: string
                                example: "Name"
                                minLength: 2
                                maxLength: 50
                              surname:
                                type: string
                                example: "Surname"
                                minLength: 2
                                maxLength: 50
                              email:
                                type: string
                                example: "name.surname@email.com"
                              status:
                                type: string
                                enum:
                                  - NEEDS_INFORMATION
                                  - ACTIVE
                              createdAt:
                                type: string
                                example: "21.04.2023"
                              role:
                                type: string
                                enum:
                                  - MENTOR
                                  - MENTEE
                              questionnaire:
                                type: object
                                properties:
                                  about:
                                    type: string
                                    minLength: 250
                                    maxLength: 1024
                                    example: "Some information about user"
                                  skills:
                                    type: array
                                    items:
                                      example: "Skill"
                                  companies:
                                    type: array
                                    items:
                                      example: "Company"
                                  rank:
                                    type: string
                                    example: "Rank"
                                  specialization:
                                    type: string
                                    example: "Specialization"
                                  hoursPerWeek:
                                    type: integer
                                    format: int32
                                    example: 20
                                    minimum: 1
                                    maximum: 168
                                  profilePhotoUrl:
                                    type: string
                                    example: "https://url.com/profile-photo.jpg"
                          to:
                            type: object
                            properties:
                              id:
                                type: integer
                                format: int64
                                example: 123
                              name:
                                type: string
                                example: "Name"
                                minLength: 2
                                maxLength: 50
                              surname:
                                type: string
                                example: "Surname"
                                minLength: 2
                                maxLength: 50
                              email:
                                type: string
                                example: "name.surname@email.com"
                              status:
                                type: string
                                enum:
                                  - NEEDS_INFORMATION
                                  - ACTIVE
                              createdAt:
                                type: string
                                example: "21.04.2023"
                              role:
                                type: string
                                enum:
                                  - MENTOR
                                  - MENTEE
                              questionnaire:
                                type: object
                                properties:
                                  about:
                                    type: string
                                    minLength: 250
                                    maxLength: 1024
                                    example: "Some information about user"
                                  skills:
                                    type: array
                                    items:
                                      example: "Skill"
                                  companies:
                                    type: array
                                    items:
                                      example: "Company"
                                  rank:
                                    type: string
                                    example: "Rank"
                                  specialization:
                                    type: string
                                    example: "Specialization"
                                  hoursPerWeek:
                                    type: integer
                                    format: int32
                                    example: 20
                                    minimum: 1
                                    maximum: 168
                                  profilePhotoUrl:
                                    type: string
                                    example: "https://url.com/profile-photo.jpg"
                          status:
                            type: string
                            enum:
                              - PENDING
                              - REJECTED
                              - ACCEPTED
                          createdAt:
                            type: string
                            example: "21.04.2023"
                          updatedAt:
                            type: string
                            example: "21.04.2023"
                      currentPageNumber:
                        type: integer
                        format: int32
                        example: 1
                      totalPages:
                        type: integer
                        format: int32
                        example: 2
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "about:blank"
                  title:
                    type: string
                    example: "Bad Request"
                  status:
                    type: integer
                    format: int32
                    example: 400
                  detail:
                    type: string
                    example: "Номер сторінки не може бути менше 1."
                  instance:
                    type: string
                    example: "/api/v1/mentoring-requests/to-me"
      security:
        - jwt_token: [ ]
  /api/v1/mentoring-requests:
    post:
      tags:
        - MentoringRequest
      summary: Creates new mentoring request from one user to another
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                toUserId:
                  type: integer
                  example: 1
      responses:
        '201':
          description: A request was successfully created
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "about:blank"
                  title:
                    type: string
                    example: "Bad Request"
                  status:
                    type: integer
                    format: int32
                    example: 400
                  detail:
                    type: string
                    example: "Користувач не заповнив профіль, тому йому не можна відправляти запити."
                  instance:
                    type: string
                    example: "/api/v1/mentoring-requests"
      security:
        - jwt_token: [ ]
  /api/v1/mentoring-requests/accept/{reqId}:
    put:
      tags:
        - MentoringRequest
      summary: Updates a mentoring request status to ACCEPTED
      parameters:
        - in: path
          name: reqId
          required: true
          example: 1
      responses:
        '204':
          description: A mentoring request status was successfully updated
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "about:blank"
                  title:
                    type: string
                    example: "Bad Request"
                  status:
                    type: integer
                    format: int32
                    example: 400
                  detail:
                    type: string
                    example: "Запит уже був опрацьований."
                  instance:
                    type: string
                    example: "/api/v1/mentoring-requests/accept/1"
        '403':
          description: User is not authorized or has not rights to accept mentoring request
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "about:blank"
                  title:
                    type: string
                    example: "Forbidden"
                  status:
                    type: integer
                    format: int32
                    example: 403
                  detail:
                    type: string
                    example: "Схоже Ви не маєте прав, щоб прийняти цей запит."
                  instance:
                    type: string
                    example: "/api/v1/mentoring-requests/accept/1"
        '404':
          description: Cannot find mentoring request by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "about:blank"
                  title:
                    type: string
                    example: "Not Found"
                  status:
                    type: integer
                    format: int32
                    example: 404
                  detail:
                    type: string
                    example: "Не вдається знайти запит, схоже його не існує."
                  instance:
                    type: string
                    example: "/api/v1/mentoring-requests/accept/1"
      security:
        - jwt_token: [ ]
  /api/v1/mentoring-requests/reject/{reqId}:
    put:
      tags:
        - MentoringRequest
      summary: Updates a mentoring request status to REJECTED
      parameters:
        - in: path
          name: reqId
          required: true
          example: 1
      responses:
        '204':
          description: A mentoring request status was successfully updated
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "about:blank"
                  title:
                    type: string
                    example: "Bad Request"
                  status:
                    type: integer
                    format: int32
                    example: 400
                  detail:
                    type: string
                    example: "Запит уже був опрацьований."
                  instance:
                    type: string
                    example: "/api/v1/mentoring-requests/reject/1"
        '403':
          description: User is not authorized or has not rights to reject mentoring request
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "about:blank"
                  title:
                    type: string
                    example: "Forbidden"
                  status:
                    type: integer
                    format: int32
                    example: 403
                  detail:
                    type: string
                    example: "Схоже Ви не маєте прав, щоб відхилити цей запит."
                  instance:
                    type: string
                    example: "/api/v1/mentoring-requests/reject/1"
        '404':
          description: Cannot find mentoring request by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "about:blank"
                  title:
                    type: string
                    example: "Not Found"
                  status:
                    type: integer
                    format: int32
                    example: 404
                  detail:
                    type: string
                    example: "Не вдається знайти запит, схоже його не існує."
                  instance:
                    type: string
                    example: "/api/v1/mentoring-requests/reject/1"
      security:
        - jwt_token: [ ]
...